Move all child components into their own folder
