+++ src/components/HorizontalBarChart/components/GradientDefs/GradientDefs.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import type {Color, GradientStop} from '../../../types';
-import {isGradientType} from '../../../utilities';
-import {LinearGradient} from '../../LinearGradient';
-import {GRADIENT_ID} from '../constants';
-import type {ColorOverrides} from '../types';
+import type {Color, GradientStop} from '../../../../types';
+import {isGradientType} from '../../../../utilities';
+import {LinearGradient} from '../../../LinearGradient';
+import {GRADIENT_ID} from '../../constants';
+import type {ColorOverrides} from '../../types';
 
 interface GradientDefsProps {
   colorOverrides: ColorOverrides[];
