+++ src/components/HorizontalBarChart/components/GradientDefs/index.ts
@@ -1,0 +1,1 @@
+export {GradientDefs, getGradientDefId} from './GradientDefs';
