+++ src/components/HorizontalBarChart/components/GroupLabel/index.ts
@@ -1,0 +1,1 @@
+export {GroupLabel} from './GroupLabel';
