+++ src/components/HorizontalBarChart/components/HorizontalBars/index.ts
@@ -1,0 +1,1 @@
+export {HorizontalBars} from './HorizontalBars';
