+++ src/components/HorizontalBarChart/components/Label/Label.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import {animated, useSpring} from '@react-spring/web';
 
-import {BARS_TRANSITION_CONFIG, FONT_SIZE} from '../../../constants';
-import {BAR_LABEL_HEIGHT} from '../constants';
+import {BARS_TRANSITION_CONFIG, FONT_SIZE} from '../../../../constants';
+import {BAR_LABEL_HEIGHT} from '../../constants';
 
 interface LabelProps {
   barHeight: number;
