+++ src/components/HorizontalBarChart/components/StackedBar/StackedBar.tsx
@@ -1,7 +1,7 @@
 import {animated, useSpring} from '@react-spring/web';
 import React from 'react';
 
-import {DataType} from '../../../types';
+import {DataType} from '../../../../types';
 
 interface StackedBarProps {
   color: string;
