+++ src/components/HorizontalBarChart/components/StackedBar/index.ts
@@ -1,0 +1,1 @@
+export {StackedBar} from './StackedBar';
