+++ src/components/HorizontalBarChart/components/StackedBars/index.ts
@@ -1,0 +1,1 @@
+export {StackedBars} from './StackedBars';
