+++ src/components/HorizontalBarChart/components/VerticalGridLines/index.ts
@@ -1,0 +1,1 @@
+export {VerticalGridLines} from './VerticalGridLines';
